More verbose error messages
Hi! Love the component.

I make the mistake of having miss-matched `value` and `children` props often and I feel the error message in its current form doesn't really assist in solving the error (I had to read the code the first time in order to solve it).

This adds a hint suggesting a resolution to the issue. Feel free to reword if needed.

Thanks 😄